tunniplaan, ehis, moodulite protokollid

+++ hois_back/src/main/java/ee/hitsa/ois/service/ehis/EhisStudentService.java
@@ -404,7 +404,7 @@ public class EhisStudentService {
         KhlKorgharidusLisa khlKorgharidusLisa = new KhlKorgharidusLisa();
 
         khlKorgharidusLisa.setOppimaAsumKp(getDate(student.getStudyStart(), directive));
-        khlKorgharidusLisa.setKursus(BigInteger.valueOf(student.getStudentGroup().getCourse()));
+        khlKorgharidusLisa.setKursus(BigInteger.valueOf(student.getStudentGroup().getCourse().longValue()));
 
         khlKorgharidusLisa.setOppekava(getCurriculum(student.getCurriculumVersion()));
         khlKorgharidusLisa.setKlOppekeel(student.getLanguage().getEhisValue());
